+++ client/src/components/MessageSent.tsx
@@ -1,4 +1,4 @@
-import { useAuthContext } from "../hooks/useAuthContext";
+import { useAuthContext } from "../hooks/auth/useAuthContext";
 import useConversation, { messageData } from "../store/useConversation";
 import { getHours } from "../utils/getHours";
