+++ client/src/hooks/auth/useAuthContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "../context/AuthContext";
+import { AuthContext } from "../../context/AuthContext";
 
 export const useAuthContext = () => {
   const context = useContext(AuthContext);
