+++ client/src/hooks/auth/useLogout.ts
