+++ client/src/hooks/messages/useSocketContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { SocketContext } from "../context/SocketContext";
+import { SocketContext } from "../../context/SocketContext";
 
 export const useSocketContext = () => {
     return useContext(SocketContext);
